big ugly push

+++ app/jobs/working_file_job.rb
@@ -7,17 +7,20 @@ class WorkingFileJob < ApplicationJob
     item = Item.find(media.item_id)
 
     FileUtils.mkdir_p(item.working_path) unless File.directory?(item.working_path)
-    FileUtils.cp(media.path, item.working_path) # copies the archive file
+    FileUtils.cp(media.path, item.working_path) unless File.exists?(item.working_path.join(media.filename))
 
-    working_file = Media.new(
-      form_id: item.form_id,
-      item_id: item.id,
-      media_type: :working,
-      filename: media.filename,
-      fieldname: media.fieldname,
-      path: item.working_path.join(media.filename)
-    )
-    working_file.file_info
+    # hash = {source_path: media.path, destination_path: item.working_path}, filename: media.filename}
+    # FileCopyJob.perform_later(hash.to_json)
+
+    working_file = Media.new(media.attributes.merge(id: nil, media_type: :working, path: item.working_path.join(media.filename)))
     working_file.save
+
+    # do we need so save the media object back to the item?
+    # item[:data][media.fieldname] << working_file
+    # item.save
+
+    ConvertingFileJob.perform_later(working_file.id)
+    # ProcessingJob.perform_later(working_file.id)
   end
+
 end
